Fix ScheduledModel objects permission check
